Added new way to handle login and logout errors, along with blog and storage

+++ backend/routes/blog/blogBucket.js
@@ -1,21 +1,22 @@
 import { Router } from "express";
 import { BlogRepository } from "./blogRepository.js";
 import multer from "multer"
+import { AppError } from "../../errors/appError.js";
 const blogImgRouter = Router();
 const storage = multer.memoryStorage();
 const upload = multer({ storage: storage });
 // Consulta sobre las imágenes del storage
 // Prueba recuperando un storage reestructurado llamado novedades
-blogImgRouter.get("/", async (req, res) => {
+blogImgRouter.get("/", async (req, res, next) => {
   try {
     const data = await BlogRepository.getAllBlogsImages();
     res.send(data);
   } catch (e) {
-    return res.status(500).json(e);
+     next(e);
   }
 });
 
-blogImgRouter.post("/", upload.single("image"), async (req, res) => {
+blogImgRouter.post("/", upload.single("image"), async (req, res, next) => {
   try {
     const { imgName, folderName } = req.body;
     const image = req.file.buffer; // El archivo cargado
@@ -23,19 +24,21 @@ blogImgRouter.post("/", upload.single("image"), async (req, res) => {
     res.send(data);
   }
   catch (err){
-    return res.status(500).json(err);
+    next(err);
   }
 })
 
 
-blogImgRouter.delete("/", async (req, res)=> {
+blogImgRouter.delete("/", async (req, res,next)=> {
   try{
     const {folderName} = req.body;
+    if(!folderName)
+      throw new AppError("MissingDataError",400,"No se enviaron los datos solicitados")
     const data = await BlogRepository.deleteImage(folderName);
     res.send(data);
   }
   catch (err){
-    res.status(500).json(err);
+    next(err);
   }
 })
